requirement.txt
ansible==2.7.9
bcrypt==3.1.7
cffi==1.13.1
cryptography==2.8
Django==2.2.6
djangorestframework==3.10.3
Jinja2==2.10.3
libvirt-python==5.7.0
MarkupSafe==1.1.1
paramiko==2.6.0
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
sqlparse==0.3.0
